Create specific table generators for headers and rows
